requirements.txt
Django==2.2.6
celery==4.3.0
PyYAML==5.1.2
